created createSupabaseServerClient function, created read session function
