+++ app/[locale]/(subpages)/(explore)/community/page.tsx
@@ -1,11 +1,10 @@
-import { cookies } from "next/headers";
 import { getTranslations, unstable_setRequestLocale } from "next-intl/server";
-import { createServerClient } from "@supabase/ssr";
 
 import { profilesOrderByArray } from "~/types/feed/OrderVariants";
 
 import { FeedWithSorting } from "~/components/Feed/FeedWithSorting";
 import { type localeTypes } from "~/i18n";
+import readUserSession from "~/lib/supabase/readUserSession";
 
 export async function generateMetadata({
   params: { locale },
@@ -25,23 +24,9 @@ export default async function CommunityPage({
 }) {
   unstable_setRequestLocale(locale);
 
-  const cookieStore = cookies();
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-      },
-    }
-  );
-
   const {
     data: { session },
-  } = await supabase.auth.getSession();
+  } = await readUserSession();
 
   return (
     <div className="container pb-12">
