+++ app/[locale]/(subpages)/book/[id]/[title]/page.tsx
@@ -1,13 +1,12 @@
-import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
 import { unstable_setRequestLocale } from "next-intl/server";
-import { createServerClient } from "@supabase/ssr";
 import { z } from "zod";
 
 import { AllReviewsButton } from "~/components/Book/AllReviewsButton";
 import { MyReview } from "~/components/Book/MyReview";
 import { CategoryLink } from "~/components/ui/CategoryLink";
 import { type localeTypes } from "~/i18n";
+import readUserSession from "~/lib/supabase/readUserSession";
 
 export default async function BookPage({
   params: { id, title, locale },
@@ -24,23 +23,9 @@ export default async function BookPage({
     notFound();
   }
 
-  const cookieStore = cookies();
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-      },
-    }
-  );
-
   const {
     data: { session },
-  } = await supabase.auth.getSession();
+  } = await readUserSession();
 
   return (
     <>
