+++ app/[locale]/(subpages)/book/[id]/[title]/reviews/page.tsx
@@ -1,7 +1,5 @@
-import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
 import { unstable_setRequestLocale } from "next-intl/server";
-import { createServerClient } from "@supabase/ssr";
 import { z } from "zod";
 
 import { reviewsOrderByArray } from "~/types/feed/OrderVariants";
@@ -10,6 +8,7 @@ import { FeedWithSorting } from "~/components/Feed/FeedWithSorting";
 import { BackCategoryLink } from "~/components/ui/BackCategoryLink";
 import { type localeTypes } from "~/i18n";
 import { db } from "~/lib/db";
+import readUserSession from "~/lib/supabase/readUserSession";
 
 export default async function BookReviewsPage({
   params: { id, title, locale },
@@ -26,23 +25,9 @@ export default async function BookReviewsPage({
     notFound();
   }
 
-  const cookieStore = cookies();
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-      },
-    }
-  );
-
   const {
     data: { session },
-  } = await supabase.auth.getSession();
+  } = await readUserSession();
 
   const reviewsQuantity = await db.review.count({
     where: {
