+++ app/[locale]/(subpages)/search/page.tsx
@@ -1,10 +1,9 @@
 import { Suspense } from "react";
-import { cookies } from "next/headers";
 import { getTranslations, unstable_setRequestLocale } from "next-intl/server";
-import { createServerClient } from "@supabase/ssr";
 
 import { SearchComponent } from "~/components/Search/SearchComponent";
 import { type localeTypes } from "~/i18n";
+import readUserSession from "~/lib/supabase/readUserSession";
 
 import Loading from "./loading";
 
@@ -26,23 +25,9 @@ export default async function SearchPage({
 }) {
   unstable_setRequestLocale(locale);
 
-  const cookieStore = cookies();
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-      },
-    }
-  );
-
   const {
     data: { session },
-  } = await supabase.auth.getSession();
+  } = await readUserSession();
 
   return (
     <Suspense fallback={<Loading />}>
