+++ app/api/auth/callback/route.ts
@@ -36,5 +36,5 @@ export async function GET(request: Request) {
   }
 
   // return the user to an error page with instructions
-  return NextResponse.redirect(`${origin}/auth/auth-code-error`);
+  return NextResponse.redirect("/login");
 }
