+++ app/api/book/manage/owned-as/route.ts
@@ -1,39 +1,15 @@
-import { cookies } from "next/headers";
-import { type CookieOptions, createServerClient } from "@supabase/ssr";
-
 import { type OwnedAsType } from "~/types/CategoryTypes";
 
 import { db } from "~/lib/db";
+import readUserSession from "~/lib/supabase/readUserSession";
 import { OwnedAsValidator } from "~/lib/validations/book/ownedAs";
 import { GlobalErrors } from "~/lib/validations/errorsEnums";
 
 export async function POST(req: Request) {
   try {
-    const cookieStore = cookies();
-
-    const supabase = createServerClient(
-      process.env.NEXT_PUBLIC_SUPABASE_URL!,
-      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-      {
-        cookies: {
-          get(name: string) {
-            return cookieStore.get(name)?.value;
-          },
-          set(name: string, value: string, options: CookieOptions) {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-            cookieStore.set({ name, value, ...options });
-          },
-          remove(name: string, options: CookieOptions) {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-            cookieStore.set({ name, value: "", ...options });
-          },
-        },
-      }
-    );
-
     const {
       data: { session },
-    } = await supabase.auth.getSession();
+    } = await readUserSession();
 
     if (!session?.user) {
       return new Response(GlobalErrors.UNAUTHORIZED);
