+++ app/api/profile/username/route.ts
@@ -1,38 +1,15 @@
-import { cookies } from "next/headers";
-import { type CookieOptions, createServerClient } from "@supabase/ssr";
 import { z } from "zod";
 
 import { db } from "~/lib/db";
+import readUserSession from "~/lib/supabase/readUserSession";
 import { UsernameValidator } from "~/lib/validations/auth";
 import { AuthErrors, GlobalErrors } from "~/lib/validations/errorsEnums";
 
 export async function PATCH(req: Request) {
   try {
-    const cookieStore = cookies();
-
-    const supabase = createServerClient(
-      process.env.NEXT_PUBLIC_SUPABASE_URL!,
-      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-      {
-        cookies: {
-          get(name: string) {
-            return cookieStore.get(name)?.value;
-          },
-          set(name: string, value: string, options: CookieOptions) {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-            cookieStore.set({ name, value, ...options });
-          },
-          remove(name: string, options: CookieOptions) {
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-            cookieStore.set({ name, value: "", ...options });
-          },
-        },
-      }
-    );
-
     const {
       data: { session },
-    } = await supabase.auth.getSession();
+    } = await readUserSession();
 
     if (!session?.user) {
       return new Response(GlobalErrors.UNAUTHORIZED);
