+++ components/Navbar/Navbar.tsx
@@ -1,29 +1,14 @@
 import type { FC } from "react";
-import { cookies } from "next/headers";
-import { createServerClient } from "@supabase/ssr";
 
 import { db } from "~/lib/db";
+import readUserSession from "~/lib/supabase/readUserSession";
 
 import { NavbarLink } from "./NavbarLink";
 
 export const Navbar: FC = async () => {
-  const cookieStore = cookies();
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-      },
-    }
-  );
-
   const {
     data: { session },
-  } = await supabase.auth.getSession();
+  } = await readUserSession();
 
   const userData =
     session?.user &&
