+++ components/Profile/FollowPage.tsx
@@ -1,11 +1,10 @@
 import type { FC } from "react";
-import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
-import { createServerClient } from "@supabase/ssr";
 
 import { profilesOrderByArray } from "~/types/feed/OrderVariants";
 
 import { db } from "~/lib/db";
+import readUserSession from "~/lib/supabase/readUserSession";
 
 import { FeedWithSorting } from "../Feed/FeedWithSorting";
 import { BackCategoryLink } from "../ui/BackCategoryLink";
@@ -19,23 +18,9 @@ export const FollowPage: FC<FollowPageProps> = async ({
   fullname,
   variant,
 }) => {
-  const cookieStore = cookies();
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-      },
-    }
-  );
-
   const {
     data: { session },
-  } = await supabase.auth.getSession();
+  } = await readUserSession();
 
   const userData = await db.profile.findUnique({
     where: {
