+++ lib/supabase/server.ts
@@ -1,0 +1,32 @@
+"use server";
+
+import { unstable_noStore } from "next/cache";
+import { cookies } from "next/headers";
+import { type CookieOptions, createServerClient } from "@supabase/ssr";
+
+// eslint-disable-next-line @typescript-eslint/require-await
+export default async function createSupabaseServerClient() {
+  unstable_noStore();
+
+  const cookieStore = cookies();
+
+  return createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        get(name: string) {
+          return cookieStore.get(name)?.value;
+        },
+        set(name: string, value: string, options: CookieOptions) {
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+          cookieStore.set({ name, value, ...options });
+        },
+        remove(name: string, options: CookieOptions) {
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+          cookieStore.set({ name, value: "", ...options });
+        },
+      },
+    }
+  );
+}
